+++ backend/canisters/local_user_index/impl/src/updates/c2c_create_group.rs
@@ -32,7 +32,7 @@ async fn c2c_create_group(args: Args) -> Response {
         prepare_ok.canister_id,
         Some(prepare_ok.local_user_index_canister_id),
         prepare_ok.canister_wasm,
-        candid::encode_one(&prepare_ok.init_canister_args).unwrap(),
+        msgpack::serialize_then_unwrap(&prepare_ok.init_canister_args),
         prepare_ok.cycles_to_use,
         on_canister_created,
     )