+++ backend/canisters/group/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Use MessagePack to serialize `init` args ([#8183](https://github.com/open-chat-labs/open-chat/pull/8183))
+
 ## [[2.0.1795](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1795-group)] - 2025-06-13
 
 ### Changed